+++ virtualapk-gradle-plugin/src/main/groovy/com.didi.virtualapk/BasePlugin.groovy
@@ -2,11 +2,10 @@ package com.didi.virtualapk
 
 import com.android.build.gradle.AppExtension
 import com.android.build.gradle.AppPlugin
-import com.android.build.gradle.internal.TaskContainerAdaptor
 import com.android.build.gradle.internal.TaskFactory
+import com.android.build.gradle.internal.TaskManager
 import com.android.build.gradle.internal.api.ApplicationVariantImpl
 import com.android.build.gradle.internal.variant.VariantFactory
-import com.android.builder.core.VariantConfiguration
 import com.android.builder.core.VariantType
 import com.didi.virtualapk.tasks.AssemblePlugin
 import com.didi.virtualapk.utils.Log
@@ -44,6 +43,14 @@ public abstract class BasePlugin implements Plugin<Project> {
     @Override
     public void apply(Project project) {
         this.project = project
+        project.ext.set(Constants.GRADLE_3_1_0, false)
+
+        try {
+            Class.forName('com.android.builder.core.VariantConfiguration')
+        } catch (Throwable e) {
+            // com.android.tools.build:gradle:3.1.0
+            project.ext.set(Constants.GRADLE_3_1_0, true)
+        }
 
         AppPlugin appPlugin = project.plugins.findPlugin(AppPlugin)
 
@@ -57,7 +64,7 @@ public abstract class BasePlugin implements Plugin<Project> {
                     Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
                         if ('preVariantWork' == method.name) {
                             checkVariantFactoryInvoked = true
-                            Log.i 'Plugin', "Evaluating VirtualApk's configurations..."
+                            Log.i 'VAPlugin', "Evaluating VirtualApk's configurations..."
                             boolean isBuildingPlugin = evaluateBuildingPlugin(appPlugin, project)
                             beforeCreateAndroidTasks(isBuildingPlugin)
                         }
@@ -69,7 +76,14 @@ public abstract class BasePlugin implements Plugin<Project> {
 
         project.extensions.create('virtualApk', VAExtention)
 
-        taskFactory = new TaskContainerAdaptor(project.tasks)
+        if (project.extensions.extraProperties.get(Constants.GRADLE_3_1_0)) {
+            TaskManager taskManager = Reflect.on(appPlugin).field('taskManager').get()
+            taskFactory = taskManager.getTaskFactory()
+        } else {
+            taskFactory = Reflect.on('com.android.build.gradle.internal.TaskContainerAdaptor')
+                    .create(project.tasks)
+                    .get()
+        }
         project.afterEvaluate {
 
             if (!checkVariantFactoryInvoked) {
@@ -84,12 +98,18 @@ public abstract class BasePlugin implements Plugin<Project> {
 
                     taskFactory.create(variantPluginTaskName, AssemblePlugin, configAction)
 
-                    taskFactory.named("assemblePlugin", new Action<Task>() {
+                    Action action = new Action<Task>() {
                         @Override
                         void execute(Task task) {
                             task.dependsOn(variantPluginTaskName)
                         }
-                    })
+                    }
+
+                    if (project.extensions.extraProperties.get(Constants.GRADLE_3_1_0)) {
+                        taskFactory.configure("assemblePlugin", action)
+                    } else {
+                        taskFactory.named("assemblePlugin", action)
+                    }
                 }
             }
         }
@@ -120,22 +140,33 @@ public abstract class BasePlugin implements Plugin<Project> {
             }
 
             appPlugin.variantManager.productFlavors.each {
-                String variantName = VariantConfiguration.computeFullName(it.key, buildType, VariantType.DEFAULT, null)
+                String variantName
+                if (project.extensions.extraProperties.get(Constants.GRADLE_3_1_0)) {
+                    variantName = Reflect.on('com.android.build.gradle.internal.core.VariantConfiguration')
+                            .call('computeFullName', it.key, buildType, VariantType.DEFAULT, null)
+                            .get()
+                } else {
+                    variantName = Reflect.on('com.android.builder.core.VariantConfiguration')
+                            .call('computeFullName', it.key, buildType, VariantType.DEFAULT, null)
+                            .get()
+                }
                 def variantPluginTaskName = createPluginTaskName("assemble${variantName.capitalize()}Plugin".toString())
                 pluginTasks.add(variantPluginTaskName)
             }
         }
 
 //        pluginTasks.each {
-//            Log.i 'Plugin', "pluginTask: ${it}"
+//            Log.i 'VAPlugin', "pluginTask: ${it}"
 //        }
 
         boolean isBuildingPlugin = false
         NameMatcher nameMatcher = new NameMatcher()
         targetTasks.every {
-            String taskName = nameMatcher.find(it, pluginTasks)
+            int index = it.lastIndexOf(":");
+            String task = index >= 0 ? it.substring(index + 1) : it
+            String taskName = nameMatcher.find(task, pluginTasks)
             if (taskName != null) {
-//                Log.i 'Plugin', "Found task name '${taskName}' by given name '${it}'"
+//                Log.i 'VAPlugin', "Found task name '${taskName}' by given name '${it}'"
                 isBuildingPlugin = true
                 return false
             }