+++ virtualapk-gradle-plugin/src/main/groovy/com.didi.virtualapk/collector/dependence/AarDependenceInfo.groovy
@@ -1,8 +1,12 @@
 package com.didi.virtualapk.collector.dependence
 
+import com.android.SdkConstants
+import com.android.build.gradle.internal.TaskManager
 import com.android.builder.model.AndroidLibrary
+import com.android.utils.FileUtils
 import com.didi.virtualapk.collector.res.ResourceEntry
 import com.didi.virtualapk.collector.res.StyleableEntry
+import com.didi.virtualapk.utils.Log
 import com.google.common.collect.ArrayListMultimap
 import com.google.common.collect.ListMultimap
 import com.google.common.collect.Lists
@@ -18,6 +22,8 @@ class AarDependenceInfo extends DependenceInfo {
      * Android library dependence in android build system, delegate of AarDependenceInfo
      */
     AndroidLibrary library
+
+    File intermediatesFile
 
     /**
      * All resources(e.g. drawable, layout...) this library can access
@@ -36,6 +42,7 @@ class AarDependenceInfo extends DependenceInfo {
 
     @Override
     File getJarFile() {
+        Log.i 'AarDependenceInfo', "Found [${library.resolvedCoordinates}]'s jar file: ${library.jarFile}"
         return library.jarFile
     }
 
@@ -45,14 +52,17 @@ class AarDependenceInfo extends DependenceInfo {
     }
     
     File getAssetsFolder() {
+        Log.i 'AarDependenceInfo', "Found [${library.resolvedCoordinates}]'s assets folder: ${library.assetsFolder}"
         return library.assetsFolder
     }
 
     File getJniFolder() {
+        Log.i 'AarDependenceInfo', "Found [${library.resolvedCoordinates}]'s jni folder: ${library.jniFolder}"
         return library.jniFolder
     }
 
     Collection<File> getLocalJars() {
+        Log.i 'AarDependenceInfo', "Found [${library.resolvedCoordinates}]'s local jars: ${library.localJars}"
         return library.localJars
     }
 
@@ -64,8 +74,9 @@ class AarDependenceInfo extends DependenceInfo {
 
         def resKeys = [] as Set<String>
 
-        def rSymbol = library.symbolFile
+        def rSymbol = getFile(library.symbolFile, TaskManager.DIR_BUNDLES, library.projectVariant, SdkConstants.FN_RESOURCE_TEXT)
         if (rSymbol.exists()) {
+            Log.i 'AarDependenceInfo', "Found [${library.resolvedCoordinates}]'s symbol file: ${rSymbol}"
             rSymbol.eachLine { line ->
                 if (!line.empty) {
                     def tokenizer = new StringTokenizer(line)
@@ -87,7 +98,41 @@ class AarDependenceInfo extends DependenceInfo {
      * @return package name of this library
      */
     public String getPackage() {
-        def xmlManifest = new XmlParser().parse(library.manifest)
+        File manifest = getFile(library.manifest, 'manifests', 'full', library.projectVariant, SdkConstants.ANDROID_MANIFEST_XML)
+        Log.i 'AarDependenceInfo', "Found [${library.resolvedCoordinates}]'s manifest file: ${manifest}"
+        def xmlManifest = new XmlParser().parse(manifest)
         return xmlManifest.@package
+    }
+
+    File getIntermediatesDir() {
+        if (intermediatesFile == null) {
+            String path = library.folder.path
+            try {
+                intermediatesFile = new File(path.substring(0, path.indexOf("${File.separator}intermediates${File.separator}")), 'intermediates')
+
+            } catch (Exception e) {
+                Log.e('AarDependenceInfo', "Can not find [${library.resolvedCoordinates}]'s intermediates dir from the path: ${path}")
+                intermediatesFile = library.folder
+            }
+        }
+        return intermediatesFile
+    }
+
+    File getFile(File defaultFile, String... paths) {
+        if (library.projectVariant == null) {
+            return defaultFile
+        }
+
+        if (defaultFile.exists()) {
+            return defaultFile
+        }
+
+        // module library
+        return FileUtils.join(intermediatesDir, paths)
+     }
+
+    @Override
+    String toString() {
+        return "${super.toString()} -> ${library}"
     }
 }