+++ virtualapk-gradle-plugin/src/main/groovy/com.didi.virtualapk/hooker/MergeManifestsHooker.groovy
@@ -3,6 +3,7 @@ package com.didi.virtualapk.hooker
 import com.android.build.gradle.api.ApkVariant
 import com.android.build.gradle.internal.scope.TaskOutputHolder
 import com.android.build.gradle.tasks.MergeManifests
+import com.didi.virtualapk.Constants
 import com.didi.virtualapk.collector.dependence.DependenceInfo
 import com.didi.virtualapk.utils.Log
 import com.didi.virtualapk.utils.Reflect
@@ -39,7 +40,7 @@ class MergeManifestsHooker extends GradleTaskHooker<MergeManifests> {
     @Override
     void beforeTaskExecute(MergeManifests task) {
 
-        def stripAarNames = virtualApk.stripDependencies.
+        def stripAarNames = vaContext.stripDependencies.
                 findAll {
                     it.dependenceType == DependenceInfo.DependenceType.AAR
                 }.
@@ -57,8 +58,17 @@ class MergeManifestsHooker extends GradleTaskHooker<MergeManifests> {
      */
     @Override
     void afterTaskExecute(MergeManifests task) {
-        variantData.outputScope.getOutputs(TaskOutputHolder.TaskOutputType.MERGED_MANIFESTS).each {
-            rewrite(it.outputFile)
+        if (project.extensions.extraProperties.get(Constants.GRADLE_3_1_0)) {
+            File outputFile = Reflect.on('com.android.build.gradle.internal.scope.ExistingBuildElements')
+                    .call('from', TaskOutputHolder.TaskOutputType.MERGED_MANIFESTS, scope.getOutput(TaskOutputHolder.TaskOutputType.MERGED_MANIFESTS))
+                    .call('element', variantData.outputScope.mainSplit)
+                    .call('getOutputFile')
+                    .get()
+            rewrite(outputFile)
+        } else {
+            variantData.outputScope.getOutputs(TaskOutputHolder.TaskOutputType.MERGED_MANIFESTS).each {
+                rewrite(it.outputFile)
+            }
         }
     }